Extract history to a new module
