+++ app/src/main/java/com/duckduckgo/app/browser/pageloadpixel/PageLoadedHandler.kt
@@ -19,13 +19,13 @@ package com.duckduckgo.app.browser.pageloadpixel
 import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.app.browser.pageloadpixel.PageLoadedSites.Companion.sites
 import com.duckduckgo.app.di.AppCoroutineScope
-import com.duckduckgo.app.history.SaveToHistory
 import com.duckduckgo.app.pixels.remoteconfig.OptimizeTrackerEvaluationRCWrapper
 import com.duckduckgo.autoconsent.api.Autoconsent
 import com.duckduckgo.browser.api.WebViewVersionProvider
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.common.utils.device.DeviceInfo
 import com.duckduckgo.di.scopes.AppScope
+import com.duckduckgo.history.api.HistoryApi
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
@@ -44,7 +44,7 @@ class RealPageLoadedHandler @Inject constructor(
     private val dispatcherProvider: DispatcherProvider,
     private val autoconsent: Autoconsent,
     private val optimizeTrackerEvaluationRCWrapper: OptimizeTrackerEvaluationRCWrapper,
-    private val saveToHistory: SaveToHistory,
+    private val historyApi: HistoryApi,
 ) : PageLoadedHandler {
 
     override fun onPageLoaded(url: String, title: String?, start: Long, end: Long) {
@@ -60,7 +60,7 @@ class RealPageLoadedHandler @Inject constructor(
                     ),
                 )
             }
-            saveToHistory.saveToHistory(url, title)
+            historyApi.saveToHistory(url, title)
         }
     }
 }
