+++ history/history-impl/src/main/java/com/duckduckgo/history/impl/store/HistoryDao.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.history.store
+package com.duckduckgo.history.impl.store
 
 import androidx.room.Dao
 import androidx.room.Insert
