+++ history/history-impl/src/main/java/com/duckduckgo/history/impl/store/HistoryDatabase.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2024 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.history.impl.store
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
+
+@Database(
+    exportSchema = true,
+    version = 1,
+    entities = [
+        HistoryEntryEntity::class,
+        VisitEntity::class,
+    ],
+)
+abstract class HistoryDatabase : RoomDatabase() {
+    abstract fun historyDao(): HistoryDao
+}
+
+val ALL_MIGRATIONS = emptyArray<Migration>()
